requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# CrawlAndParse.py: 10,11
# check_if_grade_uploaded.py: 6,7
Pillow == 4.1.1

# CrawlAndParse.py: 9
# check_if_grade_uploaded.py: 2
PyYAML == 3.12

# CrawlAndParse.py: 3
# check_if_grade_uploaded.py: 4
beautifulsoup4 == 4.6.0

# check_if_grade_uploaded.py: 12
eventlet == 0.21.0

# CrawlAndParse.py: 5
pandas == 0.20.1

# CrawlAndParse.py: 12
# check_if_grade_uploaded.py: 8
pytesseract == 0.1.7

# CrawlAndParse.py: 2
# check_if_grade_uploaded.py: 3
requests == 2.14.2
